Fix Latitude/Longitude data type (FLOAT→DECIMAL) and DateTime→DateTime2 for DB consistency

+++ migrations/2026_02_27_add_missing_ads_columns.sql
@@ -10,12 +10,12 @@ END
 
 IF COL_LENGTH('Ads', 'Latitude') IS NULL
 BEGIN
-    ALTER TABLE Ads ADD Latitude FLOAT NULL;
+    ALTER TABLE Ads ADD Latitude DECIMAL(9, 6) NULL;
 END
 
 IF COL_LENGTH('Ads', 'Longitude') IS NULL
 BEGIN
-    ALTER TABLE Ads ADD Longitude FLOAT NULL;
+    ALTER TABLE Ads ADD Longitude DECIMAL(9, 6) NULL;
 END
 
 IF COL_LENGTH('Ads', 'MainImageUrl') IS NULL
